requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
Pillow==9.3.0
pycodestyle==2.10.0
python-docx==0.8.11
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
